Spotify Audio Features Dataset

This dataset contains audio features for various songs sourced from Spotify, including metadata and audio characteristics. It represents unique song entries, with a total of 29,386 rows and 22 columns.

Usage

data(audio_features)

Format

A data frame with 29,386 rows and 22 variables:

song_id

Unique identifier for each song.

performer

Name of the artist or group performing the song.

song

Title of the song.

spotify_genre

Genre(s) of the song, represented as a list (may include multiple genres).

spotify_track_id

Unique identifier for the track in Spotify's database.

spotify_track_preview_url

URL for a preview of the track.

spotify_track_duration_ms

Duration of the track in milliseconds.

spotify_track_explicit

Logical; Indicates if the track contains explicit content (TRUE or FALSE, may be NA).

spotify_track_album

Title of the album that includes the track.

danceability

Measure of how suitable a track is for dancing, on a scale from 0.0 to 1.0.

energy

Measure of intensity and activity, on a scale from 0.0 to 1.0.

key

Key of the track, represented as an integer.

loudness

Overall loudness of a track in decibels.

mode

Modality of the track, with 1 indicating major and 0 indicating minor.

speechiness

Measure of the presence of spoken words in a track, on a scale from 0.0 to 1.0.

acousticness

Measure of the acoustic nature of a track, on a scale from 0.0 to 1.0.

instrumentalness

Measure of whether a track contains no vocals, on a scale from 0.0 to 1.0.

liveness

Measure of the presence of an audience in the recording, on a scale from 0.0 to 1.0.

valence

Measure of musical positiveness or happiness, on a scale from 0.0 to 1.0.

tempo

Tempo of the track in beats per minute.

time_signature

Time signature of the track, represented as an integer.

spotify_track_popularity

Popularity of the track on a scale from 0 to 100, where 100 is the most popular.

Source

Data sourced from the TidyTuesday project and Spotify's Web API.

Details

The dataset was cleaned to retain only one row per song (song_id), where the first occurrence of each song was selected.

Examples

data(audio_features)
head(audio_features)
#> # A tibble: 6 × 22
#>   song_id  performer song  spotify_genre spotify_track_id spotify_track_previe…¹
#>   <chr>    <chr>     <chr> <chr>         <chr>            <chr>                 
#> 1 -twisti… Bill Bla… -twi… []            NA               NA                    
#> 2 ¿Dònde … Augie Ri… ¿Dòn… ['novelty']   NA               NA                    
#> 3 ......A… Andy Wil… ....… ['adult stan… 3tvqPPpXyIgKrm4… https://p.scdn.co/mp3…
#> 4 ...And … Sandy Ne… ...A… ['rock-and-r… 1fHHq3qHU8wpRKH… NA                    
#> 5 ...Baby… Britney … ...B… ['dance pop'… 3MjUtNVVq3C8Fn0… https://p.scdn.co/mp3…
#> 6 ...Read… Taylor S… ...R… ['pop', 'pos… 2yLa0QULdQr0qAI… NA                    
#> # ℹ abbreviated name: ¹​spotify_track_preview_url
#> # ℹ 16 more variables: spotify_track_duration_ms <dbl>,
#> #   spotify_track_explicit <lgl>, spotify_track_album <chr>,
#> #   danceability <dbl>, energy <dbl>, key <dbl>, loudness <dbl>, mode <dbl>,
#> #   speechiness <dbl>, acousticness <dbl>, instrumentalness <dbl>,
#> #   liveness <dbl>, valence <dbl>, tempo <dbl>, time_signature <dbl>,
#> #   spotify_track_popularity <dbl>