Billboard Dataset

This dataset contains information about Billboard Hot 100 chart rankings over time. The data was sourced from the TidyTuesday project.

Usage

data(billboard)

Format

A data frame with 327,895 rows and 11 variables:

url

URL to the Billboard chart entry for the song

week_id

The date of the chart week

week_position

The song's position on the chart for that week

song

Title of the song

performer

Name of the artist or group performing the song

song_id

Unique identifier for the song

instance

The instance number of the song's entry on the chart

previous_week_position

The song's position on the chart in the previous week

peak_position

The highest position reached by the song on the chart

weeks_on_chart

The total number of weeks the song has been on the chart

year

The year the song was on the chart

Source

TidyTuesday Billboard Dataset

Details

The dataset was processed to convert week_id into a date format and extract the year from week_id.

Examples

data(billboard)
head(billboard)
#> # A tibble: 6 × 11
#>   url                  week_id    week_position song  performer song_id instance
#>   <chr>                <date>             <dbl> <chr> <chr>     <chr>      <dbl>
#> 1 http://www.billboar… 1965-07-17            34 Don'… Patty Du… Don't …        1
#> 2 http://www.billboar… 1965-07-24            22 Don'… Patty Du… Don't …        1
#> 3 http://www.billboar… 1965-07-31            14 Don'… Patty Du… Don't …        1
#> 4 http://www.billboar… 1965-08-07            10 Don'… Patty Du… Don't …        1
#> 5 http://www.billboar… 1965-08-14             8 Don'… Patty Du… Don't …        1
#> 6 http://www.billboar… 1965-08-21             8 Don'… Patty Du… Don't …        1
#> # ℹ 4 more variables: previous_week_position <dbl>, peak_position <dbl>,
#> #   weeks_on_chart <dbl>, year <dbl>