Feature Long Dataset

This dataset combines Billboard chart data and audio features for songs that were on the chart from 2010 to 2019. The data was sourced from the Billboard charts and audio features dataset, with entries merged based on the song_id.

Usage

data(feature_long)

Format

A data frame with 522,000 rows and 20 variables:

url

The URL for the Billboard chart for the given week.

week_id

The date of the Billboard chart for the given week.

week_position

The position of the song on the Billboard chart for the given week.

song_id

A unique identifier for the song, composed of the song name and performer.

instance

The number of times this song has appeared on the chart.

weeks_on_chart

The total number of weeks the song has spent on the chart.

year

The year the song was on the Billboard chart.

spotify_genre

A list of genres associated with the song on Spotify.

spotify_track_id

The Spotify ID for the track.

spotify_track_preview_url

The URL for the preview of the track on Spotify.

spotify_track_duration_ms

The duration of the track in milliseconds.

spotify_track_explicit

Indicates if the track contains explicit content (TRUE or FALSE).

spotify_track_album

The name of the album the track appears on.

tempo

The tempo of the track in beats per minute (BPM).

time_signature

The time signature of the track (typically 4 for most popular music).

spotify_track_popularity

The popularity score of the track on Spotify (0-100).

song

The title of the song.

performer

The performer of the song.

features

Audio features extracted from the song (e.g., danceability, energy, loudness).

value

The corresponding values of the audio features.

Source

Billboard charts and audio features dataset

Details

The dataset includes only distinct entries for each song, with audio features transformed into a long format for analysis.

Examples

data(feature_long)
head(feature_long)
#> # A tibble: 6 × 20
#>   url             week_id    week_position song_id instance weeks_on_chart  year
#>   <chr>           <date>             <dbl> <chr>      <dbl>          <dbl> <dbl>
#> 1 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> 2 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> 3 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> 4 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> 5 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> 6 http://www.bil… 2010-06-26            67 Beauti…        1              1  2010
#> # ℹ 13 more variables: spotify_genre <chr>, spotify_track_id <chr>,
#> #   spotify_track_preview_url <chr>, spotify_track_duration_ms <dbl>,
#> #   spotify_track_explicit <lgl>, spotify_track_album <chr>, tempo <dbl>,
#> #   time_signature <dbl>, spotify_track_popularity <dbl>, song <chr>,
#> #   performer <chr>, features <chr>, value <dbl>